signal_system/requirements.txt
anthropic>=0.40.0
telethon>=1.36.0
python-telegram-bot>=21.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
schedule>=1.2.0
